pyproject.toml
[project]
name = "lil-llm"
version = "0.1.0"
description = "A CLI tool for interacting with large language models via BYOK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.61.0",
    "click>=8.2.1",
    "google-generativeai>=0.8.5",
    "openai>=1.99.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]

[project.scripts]
lil-llm = "lil_llm.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
